core/deriv_client.py
import asyncio
import os
from deriv_api import DerivAPI
from dotenv import load_dotenv

load_dotenv()

class DerivClient:
    def __init__(self, app_id, api_token):
        self.app_id = app_id
        self.api_token = api_token
        self.api = None

    async def connect(self):
        if not self.app_id or not self.api_token:
            raise ValueError("App ID and API Token must be provided.")
        
        self.api = DerivAPI(app_id=self.app_id)
        
        # Authorize
        authorize = await self.api.authorize(self.api_token)
        print(f"Authorized: {authorize['authorize']['email']}")
        return self.api

    async def subscribe_ticks(self, symbol):
        source_tick = await self.api.subscribe({'ticks': symbol})
        return source_tick

    async def get_balance(self):
        response = await self.api.balance()
        return response['balance']

    async def buy_contract(self, contract_type, amount, symbol, duration, duration_unit, barrier=None):
        # This is a placeholder for buying contracts. 
        # For pending orders (Buy Stop/Sell Stop), Deriv uses a different mechanism or might not support them directly on all asset types via simple API calls in the same way MT5 does.
        # However, for this strategy, we might need to simulate pending orders or use 'proposal' and 'buy'.
        # But wait, Deriv API allows 'buy' and 'sell'. 
        # For "Pending Orders" like Buy Stop/Sell Stop on Synthetics, we usually trade CFDs (MT5) or Options (DTrader).
        # The user specified "Deriv's Synthetic Indices" and "Buy Stop/Sell Stop". 
        # If this is for DTrader (Options/Multipliers), "Pending Orders" are not standard.
        # If this is for MT5, we would use MT5.
        # BUT the architecture says "Deriv API" and "Trading Engine (Python asyncio)".
        # This implies we are trading directly on the Deriv platform (DTrader/SmartTrader equivalent) OR using the API to trade MT5 accounts (which is not directly possible via standard Deriv API, usually requires MT5 terminal).
        # HOWEVER, Deriv has a "trading" API for their proprietary platforms.
        # Let's assume we are trading "Multipliers" or "Fall/Rise" or similar, BUT "Buy Stop" implies a price level trigger.
        # If we are building a custom engine, we implement "Pending Orders" LOCALLY.
        # i.e., We watch the price, and when it hits the level, we execute a MARKET order.
        pass

    async def disconnect(self):
        if self.api:
            await self.api.disconnect()
