requirements.txt
python-deriv-api
websockets
aiohttp
python-dotenv
fastapi
uvicorn
ta-lib
schedule
python-dotenv
